# server/public/openapi.yaml
openapi: 3.1.0
info:
  title: TeamSpark Pachinko API
  version: 1.0.0
  description: |
    HTTP API for policy configuration, message and alert review, analytics,
    Arcade Engine webhooks, retention, and application settings.
servers:
  - url: http://localhost:3000/api/v1
    description: API base (Next.js app)
...
components:
  schemas:
    Meta:
      type: object
      properties:
        apiVersion:
          type: string
          enum: ["1.0"]
        status:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Status message
...
    Message:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        userId:
          type: string
        clientId:
          type: string
        sourceIP:
          type: string
        serverName:
          type: string
        sessionId:
          type: string
        messageId:
          type: string
        payloadMethod:
          type: string
        payloadToolName:
          type: string
        params:
          type: object
          nullable: true
        result:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
        requestAction:
          type: string
        rawParams:
          type: object
          nullable: true
        responseAction:
          type: string
        rawResult:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        timestampResult:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the message received a response
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            type: object
            properties:
              meta:
                $ref: '#/components/schemas/Meta'
paths:
  /messages:
    get:
      tags: [Messages]
      summary: List messages
      parameters:
        - name: sort
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: cursor
          in: query
          schema: { type: integer }
        - name: origin
          in: query
          schema: { type: string, enum: [client, server] }
        - name: userId
          in: query
          schema: { type: string }
        - name: source
          in: query
          schema: { type: string }
        - name: payloadToolkit
          in: query
          schema: { type: string }
        - name: payloadMethod
          in: query
          schema: { type: string }
        - name: payloadToolName
          in: query
          schema: { type: string }
        - name: payloadMessageId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Messages and pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: "#/components/schemas/Meta"
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
                  pagination:
                    type: object
        "500":
          $ref: "#/components/responses/ServerError"
...
  /appSettings:
    get:
      tags: [Settings]
      summary: Get application settings
      description: Returns persisted settings plus resolvedPublicBaseUrl for Arcade webhook URLs.
      responses:
        "200":
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "500":
          description: Plain JSON error object with an error string (not the usual meta envelope).
    put:
      tags: [Settings]
      summary: Update application settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filterApiBearerToken: { type: string }
                messageRetentionDays: { type: integer, minimum: 1 }
                alertRetentionDays: { type: integer, minimum: 1 }
                externalBaseUrl: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        "500":
          description: Plain JSON error object with an error string.
...